+++ client/flutter_client/lib/screens/search/search.dart
@@ -196,7 +196,7 @@ class Topics extends StatelessWidget {
             String dp = data['dp'];
             String peoples = data['peoples'].length.toString();
             bool private = data['private'];
-            double rating = data['avgRating'];
+            double rating = data['avgRating'].toDouble();
             TopicTile tile = new TopicTile(
               creator: creator,
               dp: dp,
